[SM-46] correcao quantidade de item nula no cadastro/alteracao no tipo da quantidade entrada e saida

+++ src/main/java/com/projetos/skymaster/skymastergerentesobras/controllers/registro/CadastrarSaidaController.java
@@ -164,7 +164,7 @@ public class CadastrarSaidaController {
             System.out.println("Nenhuma correspondência encontrada.");
         }
 
-        int qtdSaida = Integer.parseInt(quantidadeSaida);
+        double qtdSaida = Double.parseDouble(quantidadeSaida);
 
         RegistroDao registroDao = new RegistroDao();
         registroDao.createRegistroSaida(tipoItem, descricaoItem, nomeObra, qtdSaida, numNotaSaida, nomeUsuario);
